+++ src/visual/WaterfallCanvas.h
@@ -60,12 +60,12 @@ private:
     fftw_complex *in, *out;
     fftw_plan plan;
 
-    float fft_ceil_ma, fft_ceil_maa;
-    float fft_floor_ma, fft_floor_maa;
+    double fft_ceil_ma, fft_ceil_maa;
+    double fft_floor_ma, fft_floor_maa;
 
-    std::vector<float> fft_result;
-    std::vector<float> fft_result_ma;
-    std::vector<float> fft_result_maa;
+    std::vector<double> fft_result;
+    std::vector<double> fft_result_ma;
+    std::vector<double> fft_result_maa;
 
     WaterfallContext *glContext;
     MouseTracker mTracker;
@@ -88,7 +88,7 @@ private:
 
     bool isView;
     msresamp_crcf resampler;
-    float resample_ratio;
+    double resample_ratio;
     nco_crcf nco_shift;
     int shift_freq;
